# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keasy API",
    "description": "Keasy — data pipeline execution and cloud storage API",
    "contact": {
      "name": "Ángel Iglesias Préstamo",
      "email": "angel.iglesias.prestamo@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/jobs/{id}/discover/urls": {
      "post": {
        "tags": [
          "Discovery"
        ],
        "summary": "Sign GET URLs so the browser reads the dataset directly — the reading twin of\n[`resolve_output_urls`]. It takes the list the corpus reader enumerated and\nderives none.",
        "operationId": "resolve_discover_urls",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetUrlsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed GET URLs for the requested dataset keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "400": {
            "description": "The destination connection is gone, or a path outside the dataset"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/v1/jobs/{id}/output/urls": {
      "post": {
        "tags": [
          "Discovery"
        ],
        "summary": "Sign PUT URLs so the browser uploads the output it just produced straight to\nthe job's sink.",
        "operationId": "resolve_output_urls",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetUrlsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed PUT URLs for the output keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "400": {
            "description": "The destination connection is gone, or a path outside the dataset"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetUrlsRequest": {
        "type": "object",
        "required": [
          "paths"
        ],
        "properties": {
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Paths relative to the dataset (or connection). The caller names them —\nthe executor's output, the corpus reader's enumeration — and keasy signs\nthe list it is handed."
          }
        }
      },
...
      "ResolveResponse": {
        "type": "object",
        "required": [
          "files"
        ],
        "properties": {
          "files": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}